requirements.txt
flask-bootstrap==3.3.7.1
frozen-flask==0.18
pytest==7.2.0
